fix: export errors
