+++ libs/wallet-client/template/index.ts
@@ -1,4 +1,5 @@
 export { Transaction, TransactionModel } from '@vegaprotocol/transactions'
 
 export { WalletClient } from './client'
+export { WalletClientError, WalletHttpError } from './errors'
 export { WalletModel, Identifier, WalletClientHandler, ResponseError } from './model'
